requirements.txt
numpy
pandas
setuptools
logging
google-cloud-storage
pyyaml
torch
seaborn
matplotlib
wordcloud
